requirements.txt
langchain
sentence-transformers
faiss-cpu
numpy
transformers
safetensors
torch